pkg/transport: introduce Transport interface; move noisenet to transport/raw
This is the architectural foundation for adding NoiseSocket, BOLT-8, and
any future Noise-based wire protocol without forking the codebase. The
noise cryptography (provided by flynn/noise) is identical across
transports; only the framing differs. Each transport implementation can
now plug into the noisecat runtime via a small Transport interface.

Layout:

  pkg/transport/transport.go     - Transport interface, Options struct
  pkg/transport/raw/             - the historical noisecat framing
  pkg/transport/raw/conn.go      - moved verbatim from pkg/noisenet/conn.go
  pkg/transport/raw/conn_test.go - moved tests
  pkg/transport/raw/raw.go       - Transport implementation
  pkg/noisenet/conn.go           - backwards-compat shim that re-exports
                                   pkg/transport/raw via type aliases and
                                   deprecated function wrappers

Behavior is unchanged. Existing tests still pass under -race. External
callers that imported pkg/noisenet keep compiling (the shim) but will
see Go's `Deprecated:` doc hint nudging them toward pkg/transport/raw.

Internal callers (pkg/noisecat and its tests) now import the new
pkg/transport/raw directly so staticcheck does not flag the deprecation
in our own code.

This commit deliberately makes no behavior change. The pluggable
dispatch logic (selecting a transport from a CLI flag) comes in a
follow-up so it can be reviewed alongside the first non-raw transport.
