+++ pkg/noisecat/net.go
@@ -9,7 +9,7 @@ import (
 	"syscall"
 
 	"github.com/flynn/noise"
-	"github.com/gedigi/noisecat/pkg/noisenet"
+	"github.com/gedigi/noisecat/pkg/transport/raw"
 )
 
 // Noisecat defines the main network configuration
@@ -24,7 +24,7 @@ func (n *Noisecat) StartClient() {
 	netAddress := net.JoinHostPort(n.Config.DstHost, n.Config.DstPort)
 	localAddress := net.JoinHostPort(n.Config.SrcHost, n.Config.SrcPort)
 
-	conn, err := noisenet.Dial("tcp", netAddress, localAddress, n.NoiseConfig)
+	conn, err := raw.Dial("tcp", netAddress, localAddress, n.NoiseConfig)
 	if err != nil {
 		n.Log.Fatalf("can't connect to %s/tcp: %s", netAddress, err)
 	}
@@ -39,7 +39,7 @@ func (n *Noisecat) StartClient() {
 func (n *Noisecat) StartServer() {
 	netAddress := net.JoinHostPort(n.Config.SrcHost, n.Config.SrcPort)
 
-	listener, err := noisenet.Listen("tcp", netAddress, n.NoiseConfig)
+	listener, err := raw.Listen("tcp", netAddress, n.NoiseConfig)
 	if err != nil {
 		n.Log.Fatalf("can't listen: %s", err)
 	}
